add checkValidName test in commonlib
